+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Changes to be tagged
 
+## 5.36.2 - 2020-01-26
+### Added
+- `phases` to raw quest.
+- `beforeActionVals` to `mstQuest`.
+
 ## 5.36.1 - 2020-01-23
 ### Fixed
 - Some servants should have multiple Valentine CEs.